Simple Calculator with Match Case

+++ control-flow/match_case_calculator.py
@@ -1,6 +1,6 @@
 
-num1 = float(input("Enter the first number:"))
-num2 = float(input("Enter the second number:"))
+num1 = int(input("Enter the first number:"))
+num2 = int(input("Enter the second number:"))
 operation = input ("Choose the operation (+, -, *, /): ")
 
 def calculate(num1, num2, operation):
@@ -17,3 +17,6 @@ def calculate(num1, num2, operation):
 result = calculate(num1, num2, operation)
 print(f'The result is [result]')
         
+
+
+
